Added type specs
Typespecs provide a way of compensating for the lack of a static type system. This can be
useful in conjunction with the dialyzer tool to perform static analysis of your programs.
Furthermore, typespecs allow you to better-document your functions. Remember that Elixir
is a dynamic language, so function inputs and output can’t be easily deduced by looking at
the function’s signature